Add files via upload
Portfolio optimization and analysis using Riskfolio and the quantstats library. Pulling data from Yahoo Finance

Riskfolio-Lib is a library for making portfolio optimization and quantitative strategic asset allocation in Python made in Peru 🇵🇪. Its objective is to help students, academics and practitioners to build investment portfolios based on mathematically complex models with low effort. It is built on top of CVXPY and closely integrated with pandas data structures.